Added PassportJS w/ GitHub strategy + optimization
GENERAL
-Added comments for clarity

SECURITY
-Users may now sign in using their GitHub accounts.
-Multer now checks file mimetype instead of extension to reject non-images.

STYLING
-.logout in CSS now positioned w/ flexbox instead of position prop.

ERROR HANDLING
-Error handler will now throw unexpected errors instead of sending err as json.
-Fixed bug where user image was deleted b/c fs treated JPG and jpg
---as equivalent, but deleteUserImage did not.

+++ app.js
@@ -6,6 +6,38 @@ const helmet = require('helmet');
 const cookieParser = require('cookie-parser');
 const multer = require('multer');
 
+app.use(helmet());
+
+// ### Passport Setup ###
+const session = require('express-session');
+app.use(session({
+    secret: 'qwe/.,e',
+    resave: false,
+    saveUninitialized: true
+  }))
+const passport = require('passport');
+const GitHubStrategy = require('passport-github').Strategy;
+app.use(passport.initialize());
+app.use(passport.session());
+const passportConfig = require('./config');
+
+passport.use(new GitHubStrategy(passportConfig,
+  function(accessToken, refreshToken, profile, cb) {
+    // User.findOrCreate({ githubId: profile.id }, function (err, user) {
+    //   return cb(err, user);
+    // });
+    return cb(null, profile);
+  }
+));
+passport.serializeUser((user, cb) => {
+    return cb(null, user);
+});
+passport.deserializeUser((user, cb) => {
+    return cb(null, user);
+})
+//
+
+// ### Multer Setup ###
 function getExtension(fileName){
     const dotIndex = Array.prototype.lastIndexOf.call(fileName, '.');
     return fileName.split('').slice(dotIndex).join('');
@@ -24,9 +56,7 @@ const storage = multer.diskStorage({
 const upload = multer({
     storage,
     fileFilter: (req, file, cb) => {
-        const extension = getExtension(file.originalname);
-        const acceptedExtensions = ['.png', '.jpg', '.jpeg', '.tif', '.tiff', '.JPG', '.bmp'];
-        if (!(acceptedExtensions.filter(ext => ext === extension).length > 0)) {
+        if(file.mimetype.search('image')) {
             return cb(new Error('illegalFileType'));
         }
         return cb(null, true);
@@ -35,19 +65,22 @@ const upload = multer({
         next(err);
     }
 })
+//
 
+// ### Deletes previous userImage after new upload ###
 function deleteUserImage(req, res, next){
     const acceptedExtensions = ['.png', '.jpg', '.jpeg', '.tif', '.tiff', '.JPG', '.bmp'];
     acceptedExtensions.forEach(char => {
-        if(fs.existsSync(`./public/uploads/${req.cookies.username+char}`)){
-            fs.unlinkSync(`./public/uploads/${req.cookies.username+char}`);
+        if(fs.existsSync(`./public/uploads/${req.cookies.username+char}`) 
+            && req.file.filename !== req.cookies.username+(char).toLowerCase()){ //Since fs sees .JPG and .jpg as equivalent
+                fs.unlinkSync(`./public/uploads/${req.cookies.username+char}`);
         }
     })
 
     next();
 }
+//
 
-app.use(helmet());
 app.use(express.json());
 app.use(express.urlencoded({
     extended: true
@@ -58,6 +91,7 @@ app.use(express.static('public'));
 app.set('view engine', 'ejs')
 app.set('views', path.join(__dirname, 'views'));
 
+// ### Respond to expected errors ###
 app.use((req, res, next) => {
     switch(req.query.status){
         case 'failedLogin':
@@ -67,33 +101,38 @@ app.use((req, res, next) => {
             res.locals.msg = "Please login before visiting that page.";
             break;
         case 'illegalFileType':
-            res.locals.msg = "Only .png, .jpg, .jpeg, .JPG, .tif, .tiff, and .bmp files are supported";
+            res.locals.msg = "Uploaded file was not an image!";
             break;
         default:
             res.locals.msg = '';
     }
     next();
 })
+//
 
+//### Login routing ###
 app.get('/', (req, res, next) => {
     res.redirect(req.cookies.username ? '/welcome' : '/login');
 })
 app.get('/login', (req, res, next) => {
     res.render('login');
 })
-app.get('/welcome', (req, res, next) => {
+
+app.get('/auth', 
+  passport.authenticate('github', { 
+    failureRedirect: '/login?status=failedLogin', 
+    successRedirect: '/process_github-login'
+}));
+
+app.get('/process_github-login', (req, res, next) => {
+    res.cookie('username', req.user.username);
     const users = JSON.parse(fs.readFileSync('./db-mock/db.json'));
-    const targetUser = users.find(char => char.username === req.cookies.username);
-    console.log('TARGETUSER', targetUser);
-    if(req.cookies.username){
-        res.render('welcome', {
-            username: req.cookies.username,
-            image: `<img src="${'/uploads/' + targetUser.fileName}"/>`
-        });
-    } else {
-        next(new Error('illegalAccessAttempt'));
+    if(!users.find(char => char.username === req.user.username)){
+        users.push({username: req.user.username, fileName: null});
+        fs.writeFileSync('./db-mock/db.json', JSON.stringify(users));
     }
-})
+    res.redirect('/welcome');
+});
 
 app.post('/process_login', (req, res, next) => {
     const db = JSON.parse(fs.readFileSync('./db-mock/db.json'))
@@ -105,9 +144,24 @@ app.post('/process_login', (req, res, next) => {
     } else {
         next(new Error('failedLogin'));
     }
+});
+//
+
+//### Main page routing ###
+app.get('/welcome', (req, res, next) => {
+    const users = JSON.parse(fs.readFileSync('./db-mock/db.json'));
+    const targetUser = users.find(char => char.username === req.cookies.username);
+    if(req.cookies.username){
+        res.render('welcome', {
+            username: req.cookies.username,
+            image: `<img src="${'/uploads/' + (targetUser.fileName || 'default.png')}"/>`
+        });
+    } else {
+        next(new Error('illegalAccessAttempt'));
+    }
 })
 
-app.post('/process_upload-image', deleteUserImage, upload.single('user-image'), (req, res, next) => {
+app.post('/process_upload-image', upload.single('userImage'), deleteUserImage, (req, res, next) => {
     const users = JSON.parse(fs.readFileSync('./db-mock/db.json'));
     users.map((char, i) => {
         if(char.username === req.cookies.username){
@@ -116,13 +170,15 @@ app.post('/process_upload-image', deleteUserImage, upload.single('user-image'), 
     })
     fs.writeFileSync('./db-mock/db.json', JSON.stringify(users));
     res.redirect('/welcome');
-})
+});
 
 app.get('/logout', (req, res, next) => {
     res.clearCookie('username');
     res.redirect('/login');
-})
+});
+//
 
+// ### Error Handling ###
 app.use((err, req, res, next) => {
     const expectedErrs = ['failedLogin', 'illegalAccessAttempt', 'illegalFileType'];
     if(expectedErrs.filter(error => error === err.message).length > 0){
@@ -132,12 +188,13 @@ app.use((err, req, res, next) => {
             res.redirect(`/login?status=${err.message}`);
         }
     } else { 
-        res.status(500).send(err);
+        throw err;
     }
 })
 
 app.use(function (req, res, next) {
     res.status(404).send("404 -- PAGE NOTE FOUND")
 })
+//
 
 app.listen(3000);
